+++ server/src/main/java/org/opensearch/cluster/awarenesshealth/ClusterAwarenessHealth.java
@@ -16,17 +16,25 @@ import org.opensearch.common.io.stream.StreamOutput;
 import org.opensearch.common.io.stream.Writeable;
 import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.Settings;
+import org.opensearch.common.xcontent.ToXContentFragment;
+import org.opensearch.common.xcontent.XContentBuilder;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 /**
  * Cluster state Awareness health information
  */
-public class ClusterAwarenessHealth implements Writeable {
+public class ClusterAwarenessHealth implements Writeable, ToXContentFragment, Iterable<ClusterAwarenessAttributesHealth> {
 
-    private final ClusterAwarenessAttributesHealth clusterAwarenessAttributesHealth;
+    private static final String AWARENESS_ATTRIBUTE = "awareness_attributes";
+    private final Map<String, ClusterAwarenessAttributesHealth> clusterAwarenessAttributesHealthMap;
 
     /**
      * Creates cluster awareness health from cluster state.
@@ -37,20 +45,47 @@ public class ClusterAwarenessHealth implements Writeable {
      */
     public ClusterAwarenessHealth(ClusterState clusterState, ClusterSettings clusterSettings, String awarenessAttributeName) {
         // This property will govern if we need to show unassigned shard info or not
-        boolean displayUnassignedShardLevelInfo = canCalcUnassignedShards(clusterSettings, awarenessAttributeName);
-        clusterAwarenessAttributesHealth = new ClusterAwarenessAttributesHealth(
-            awarenessAttributeName,
-            displayUnassignedShardLevelInfo,
-            clusterState
-        );
+        boolean displayUnassignedShardLevelInfo;
+        ClusterAwarenessAttributesHealth clusterAwarenessAttributesHealth;
+        clusterAwarenessAttributesHealthMap = new HashMap<>();
+        List<String> awarenessAttributeList = getAwarenessAttributeList(awarenessAttributeName, clusterSettings);
+        for (String awarenessAttribute : awarenessAttributeList) {
+            displayUnassignedShardLevelInfo = canCalcUnassignedShards(clusterSettings, awarenessAttribute);
+            clusterAwarenessAttributesHealth = new ClusterAwarenessAttributesHealth(
+                awarenessAttribute,
+                displayUnassignedShardLevelInfo,
+                clusterState
+            );
+            clusterAwarenessAttributesHealthMap.put(awarenessAttribute, clusterAwarenessAttributesHealth);
+        }
     }
 
     public ClusterAwarenessHealth(final StreamInput in) throws IOException {
-        clusterAwarenessAttributesHealth = new ClusterAwarenessAttributesHealth(in);
+        int size = in.readVInt();
+        if (size > 0) {
+            clusterAwarenessAttributesHealthMap = new HashMap<>(size);
+            for (int i = 0; i < size; i++) {
+                ClusterAwarenessAttributesHealth clusterAwarenessAttributesHealth = new ClusterAwarenessAttributesHealth(in);
+                clusterAwarenessAttributesHealthMap.put(
+                    clusterAwarenessAttributesHealth.getAwarenessAttributeName(),
+                    clusterAwarenessAttributesHealth
+                );
+            }
+        } else {
+            clusterAwarenessAttributesHealthMap = Collections.emptyMap();
+        }
     }
 
-    public ClusterAwarenessHealth(String awarenessAttribute) {
-        this.clusterAwarenessAttributesHealth = new ClusterAwarenessAttributesHealth(awarenessAttribute, Collections.emptyMap());
+    private List<String> getAwarenessAttributeList(String awarenessAttributeName, ClusterSettings clusterSettings) {
+        // Helper function to check if we need health for all or for one awareness attribute.
+        boolean displayAllAwarenessAttribute = awarenessAttributeName == null || awarenessAttributeName.isBlank();
+        List<String> awarenessAttributeList = new ArrayList<>();
+        if (!displayAllAwarenessAttribute) {
+            awarenessAttributeList.add(awarenessAttributeName);
+        } else {
+            awarenessAttributeList = clusterSettings.get(AwarenessAllocationDecider.CLUSTER_ROUTING_ALLOCATION_AWARENESS_ATTRIBUTE_SETTING);
+        }
+        return awarenessAttributeList;
     }
 
     private boolean canCalcUnassignedShards(ClusterSettings clusterSettings, String awarenessAttributeName) {
@@ -73,22 +108,26 @@ public class ClusterAwarenessHealth implements Writeable {
         return allocationAwarenessBalance && forcedZoneSettingsExists;
     }
 
-    public ClusterAwarenessAttributesHealth getAwarenessAttributeHealth() {
-        return this.clusterAwarenessAttributesHealth;
+    public Map<String, ClusterAwarenessAttributesHealth> getClusterAwarenessAttributesHealthMap() {
+        return clusterAwarenessAttributesHealthMap;
     }
 
     @Override
     public void writeTo(final StreamOutput out) throws IOException {
-        clusterAwarenessAttributesHealth.writeTo(out);
+        int size = clusterAwarenessAttributesHealthMap.size();
+        out.writeVInt(size);
+        if (size > 0) {
+            for (ClusterAwarenessAttributesHealth awarenessAttributeValueHealth : this) {
+                awarenessAttributeValueHealth.writeTo(out);
+            }
+        }
     }
 
     @Override
     public String toString() {
-        return "ClusterStateHealth{"
-            + ", clusterAwarenessAttributeHealth.awarenessAttributeName"
-            + (clusterAwarenessAttributesHealth == null ? "null" : clusterAwarenessAttributesHealth.getAwarenessAttributeName())
-            + ", clusterAwarenessAttributeHealth.size="
-            + (clusterAwarenessAttributesHealth == null ? "null" : clusterAwarenessAttributesHealth.getAwarenessAttributeHealthMap().size())
+        return "ClusterAwarenessHealth{"
+            + "clusterAwarenessHealth.clusterAwarenessAttributesHealthMap.size="
+            + (clusterAwarenessAttributesHealthMap == null ? "null" : clusterAwarenessAttributesHealthMap.size())
             + '}';
     }
 
@@ -97,11 +136,26 @@ public class ClusterAwarenessHealth implements Writeable {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         ClusterAwarenessHealth that = (ClusterAwarenessHealth) o;
-        return Objects.equals(clusterAwarenessAttributesHealth, that.clusterAwarenessAttributesHealth);
+        return clusterAwarenessAttributesHealthMap.size() == that.clusterAwarenessAttributesHealthMap.size();
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(clusterAwarenessAttributesHealth);
+        return Objects.hash(clusterAwarenessAttributesHealthMap);
+    }
+
+    @Override
+    public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
+        builder.startObject(AWARENESS_ATTRIBUTE);
+        for (ClusterAwarenessAttributesHealth awarenessAttributeValueHealth : this) {
+            awarenessAttributeValueHealth.toXContent(builder, params);
+        }
+        builder.endObject();
+        return builder;
+    }
+
+    @Override
+    public Iterator<ClusterAwarenessAttributesHealth> iterator() {
+        return clusterAwarenessAttributesHealthMap.values().iterator();
     }
 }